+++ src/typings/logs.d.ts
